+++ project/settings/__init__.py
@@ -258,6 +258,14 @@ CONSTANCE_ADDITIONAL_FIELDS = {
         {"widget": "tinymce.widgets.TinyMCE", "required": False},
     ],
     "json_field": ["django.forms.JSONField", {"required": False}],
+    "choice_location_provider_field": [
+        "django.forms.fields.ChoiceField",
+        {
+            "widget": "django.forms.Select",
+            "choices": (("nominatim", "Nominatim"),),
+        },
+    ],
+    "float_field": ["django.forms.FloatField", {"required": False}],
 }
 CONSTANCE_CONFIG = {
     "INSTANCE_TITLE": ("TerraVisu", _("Instance title"), str),
@@ -302,6 +310,46 @@ CONSTANCE_CONFIG = {
         _("Define custom MapboxDraw style for the distance measure control"),
         "json_field",
     ),
+    "SEARCH_IN_LOCATIONS": (
+        False,
+        _("Enable search in locations in the search bar map control"),
+        bool,
+    ),
+    "SEARCH_IN_LOCATIONS_PROVIDER": (
+        "Nominatim",
+        _("Search provider for the search in locations feature"),
+        "choice_location_provider_field",
+    ),
+    "NOMINATIM_URL": (
+        "https://nominatim.openstreetmap.org/search.php",
+        _("Nominatim search service base url"),
+        str,
+    ),
+    "NOMINATIM_USE_VIEWBOX": (
+        False,
+        _("Use nominatim's viewbox option to filter search results"),
+        bool,
+    ),
+    "NOMINATIM_VIEWBOX_MIN_LAT": (
+        None,
+        _("Minimum latitude coordinate for nominatim viewbox"),
+        "float_field",
+    ),
+    "NOMINATIM_VIEWBOX_MIN_LONG": (
+        None,
+        _("Minimum longitude coordinate for nominatim viewbox"),
+        "float_field",
+    ),
+    "NOMINATIM_VIEWBOX_MAX_LAT": (
+        None,
+        _("Maximum latitude coordinate for nominatim viewbox"),
+        "float_field",
+    ),
+    "NOMINATIM_VIEWBOX_MAX_LONG": (
+        None,
+        _("Maximum longitude coordinate for nominatim viewbox"),
+        "float_field",
+    ),
 }
 
 CONSTANCE_CONFIG_FIELDSETS = {
@@ -345,6 +393,14 @@ CONSTANCE_CONFIG_FIELDSETS = {
             "OPENID_DEFAULT_LOGIN_BUTTON_TEXT",
             "MEASURE_CONTROL",
             "MEASURE_DRAW_STYLES",
+            "SEARCH_IN_LOCATIONS",
+            "SEARCH_IN_LOCATIONS_PROVIDER",
+            "NOMINATIM_URL",
+            "NOMINATIM_USE_VIEWBOX",
+            "NOMINATIM_VIEWBOX_MIN_LAT",
+            "NOMINATIM_VIEWBOX_MIN_LONG",
+            "NOMINATIM_VIEWBOX_MAX_LAT",
+            "NOMINATIM_VIEWBOX_MAX_LONG",
         ),
         "collapse": True,
     },